requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.4
bcrypt==4.0.1
billiard==3.6.3.0
cached-property==1.5.2
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.4
daphne==2.5.0
dataclasses==0.8
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
Django==3.1
django-allauth==0.52.0
django-celery-beat==2.0.0
django-cors-headers==3.10.1
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==21.1
django-nine==0.2.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-restframework==0.0.1
django-timezone-field==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
flower==0.9.5
hiredis==2.0.0
humanize==2.6.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
incremental==21.3.0
jsonschema==3.2.0
kombu==4.6.11
lxml==4.8.0
Markdown==3.3.7
mongoengine==0.24.0
msgpack==0.6.2
oauthlib==3.2.2
packaging==21.3
paramiko==2.12.0
Pillow==8.4.0
pprintpp==0.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.28
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.3
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
stripe==5.1.1
texttable==1.6.7
toml==0.10.1
tornado==6.0.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.13
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.59.0
whitenoise==5.3.0
wrapt==1.14.0
zipp==3.6.0
zope.interface==5.4.0